refactor(api): split views into submodules
